apalis-tk1: lvds-tegra124: configure lvds in extended mode by default
Disabling LVDS output by default has issue with pcmanfm consuming 100% CPU if
HDMI is inactive. So to make sure HDMI is set as primiary display and to address
the issue with pcmanfm, configure LVDS in extended mode by default so that if
HDMI is not active LVDS will be on atleast.

Signed-off-by: Bhuvanchandra DV <bhuvanchandra.dv@toradex.com>
Acked-by: Marcel Ziswiler <marcel.ziswiler@toradex.com>