lxterminal: address garbled title bar with gtk+3
The old fix was no longer working after the update to gtk+3.
Add a similar fix for gtk+3 as the one we used with gtk+2.

While at it do it in a two step process, only if the original call
does return with 'NULL' use a less specific gdk call.

Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>
Acked-by: Marcel Ziswiler <marcel.ziswiler@toradex.com>