tdx-nv-binaries: add x11 dependencies to help dependent packages
As this recipe adds only binary drivers it does not need any DEPENDS
to build.
However adding depends prevents QA warnings which test the binaries
for dependencies and, more importantly other recipe will get needed
dependencies installed in their sysroot.
e.g. lxqt-panel would fail because libdamage is not installed in its recipe
specific sysroot.

While at it use -p with mkdir to not fail even if the directory already exists
from a previous (failed) run of bitbake.

While at it move the libpcre* symlinks to the nv-gstapps packages where they
are used and RDEPEND on the real libpcre* libraries.

Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>
Acked-by: Marcel Ziswiler <marcel.ziswiler@toradex.com>