glibc: Check /etc/ld.so.conf.d/*.conf by default
The expected modern behavior for dealing with adding ld.so.conf entries
is to add a file to /etc/ld.so.conf.d/.  In order to do this, ld.so.conf
needs to explicitly include that /etc/ld.so.conf.d/*.conf.  Make it so.

Signed-off-by: Tom Rini <trini@konsulko.com>
---
This has been posted to oe-core and is currently awaiting comments or
application.